get_all_subgroups renamed to get_subsets
